refactor: move tool definitions to a shared folder
