+++ src/customJs/tools/getSocialShareToolDefinition.js
