+++ src/customJs/tools/getSocialShareToolDefinition.test.js
